+++ src/conftest.py
@@ -3,3 +3,8 @@
 import os
 
 os.environ["PANDAS_CONTRACT_MODE"] = "raise"
+
+# Run initialization code for pandas_contract
+import pandas_contract
+
+del pandas_contract