+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Planned]
 
 ## [Upcoming]
+## [0.11.0]
+### Changed
+- Move _lib.ValidateDictT to _decorator
+- Use pandera.pandas imports instead of pandera directly.
 
 ## [0.10.0]
 ### Changed